requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==0.29.0
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
eventlet==0.33.3
